requirements.txt
# general modules
django==4.2
python-decouple
pillow
djangorestframework
djangorestframework-simplejwt

# third party modules
markdown       
django-filter  
django-coreapi
drf-yasg[validation]
djangorestframework-simplejwt
djoser
django-coreapi
drf-yasg[validation]
djangorestframework-simplejwt
djoser
drf_spectacular


# development modules

# Email third party modules
django-mail-templated

# refactoring and test
black
flake8
pytest
pytest-django

